+++ Frontend/src/Layout.jsx
@@ -1,15 +1,12 @@
-// import React from 'react'
-// import { Link } from 'react-router-dom'
-import NavBar from './Pages/NavBar/NavBar'
-import Footer from './Pages/Footer/Footer'
+import NavBar from './Pages/Header'
+import Footer from './Pages/Footer'
 import { Outlet,ScrollRestoration } from 'react-router-dom'
-import Register from './Pages/Registration/Register'
+// import Register from './Pages/Register'
 
 function Layout() {
   return (
     <div className='scroll-smooth'>
       <ScrollRestoration />
-      
       <NavBar/>
       <Outlet/>
       <Footer/>
