+++ Frontend/src/Pages/DonateBooks.jsx
@@ -1,5 +1,5 @@
-import { useSmoothScroll } from '../../../CustomHooks/Navigation';
-import './Books.css';
+import { useSmoothScroll } from './Navigation';
+import '../css/form.css';
 
 const currentYear = new Date().getFullYear();
