+++ Frontend/src/Pages/Footer.jsx
@@ -1,7 +1,6 @@
-// import React from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faInstagram } from '@fortawesome/free-brands-svg-icons'; // for brand icons
-import Sies_logo from './assets/sies white logo.png'
+import Sies_logo from '../assets/sies_logo_footer.png'
 import { faFacebook } from '@fortawesome/free-brands-svg-icons';
 import { faLinkedin } from '@fortawesome/free-brands-svg-icons';
 import { faYoutube } from '@fortawesome/free-brands-svg-icons';
