+++ Frontend/src/Pages/QuickLinks.jsx
@@ -1,9 +1,7 @@
 import { useState } from 'react';
-// import { Link } from 'react-router-dom'
-// import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import '../Books/Books.css'
-import {useScrollToHash} from '../../../CustomHooks/Navigation'
-import { useSmoothScroll } from '../../../CustomHooks/Navigation';
+import '../css/form.css'
+import {useScrollToHash} from './Navigation'
+import { useSmoothScroll } from './Navigation';
 
 const createOption = (value, text) => {
     return (
