+++ Frontend/src/assets/event_1.jpg
