+++ Frontend/src/assets/event_2.jpg
