+++ Frontend/src/assets/library-view.png
