+++ Frontend/src/assets/sies_logo_footer.png
