+++ Frontend/src/assets/sies_logo_header.png
