+++ Frontend/src/assets/user_icon.png
